Ensured there are no hardcoded fp16 or fp32 types, dtypes to be specified in each allocation call, or by setting default_dtype in GPU backend

+++ neon/transforms/cross_entropy.py
@@ -73,7 +73,7 @@ def cross_entropy_multi(backend, outputs, targets, temp, epsilon=2**-23,
     backend.log(temp[1], out=temp[1])
     backend.multiply(targets, temp[1], out=temp[1])
     backend.multiply(temp[1], -1.0, out=temp[0])
-    result = backend.empty((1, 1))
+    result = backend.empty((1, 1), dtype=outputs.dtype)
     if scale_by_batchsize:
         backend.divide(temp[0], temp[0].shape[1], temp[0])
     return backend.sum(temp[0], axes=None, out=result)
