+++ neon/transforms/softmax.py
@@ -54,6 +54,6 @@ class Softmax(Activation):
         self.apply_function(backend, inputs, outputs)
 
         if not self.tmp or self.tmp.shape != inputs.shape:
-            self.tmp = backend.ones(inputs.shape)
+            self.tmp = backend.ones(inputs.shape, dtype=inputs.dtype)
         if not self.skip_derivative:
             backend.softmax_gradient(outputs, err=self.tmp, out=inputs)
